fix incorrect lighting pos-processing on block that were changed
Syncing a chunk is now possible from the creative dimension itself. Make sure you are not in a place that is going to have blocks later!

config file is copied from the mod instead of created, this makes the code clearer by not having the whole config file creation into it.

No sound will be played when teleporting to the creative dimension. This has been done pretty poorly and will need more work late but is Server side only, keeping the mod only server required.